chore: remove footer section
