+++ packages/gdsc-dju-admin/src/components/atoms/Button/index.tsx
@@ -9,37 +9,20 @@ interface ButtonProps {
   disable?: boolean;
   onClick?: (e?: any) => void;
   type?: 'button' | 'submit' | 'reset';
+  size?: 'small' | 'medium' | 'large';
 }
-const GDSCButtonL: React.FC<ButtonProps> = ({
-  text,
-  color,
-  disable,
-  onClick,
-  type,
-}) => {
-  return (
-    <StyledButton
-      color={color ? 'white' : 'grey900'}
-      border={color ? color : 'white'}
-      size="large"
-      background={color ? color : 'white'}
-      disable={disable}
-      onClick={onClick}
-      type={type}
-    >
-      {text}
-    </StyledButton>
-  );
-};
+
 const GDSCButton: React.FC<ButtonProps> = ({
   text,
   color,
   disable,
   onClick,
   type,
+  size = 'medium',
 }) => {
   return (
     <StyledButton
+      size={size}
       color={color ? 'white' : 'grey900'}
       border={color ? color : 'white'}
       background={color ? color : 'white'}
@@ -52,4 +35,4 @@ const GDSCButton: React.FC<ButtonProps> = ({
   );
 };
 
-export { GDSCButtonL, GDSCButton };
+export { GDSCButton };
