+++ packages/gdsc-dju-admin/src/components/atoms/Button/styled.ts
@@ -13,7 +13,7 @@ export const StyledButton = styled.button<{
   justify-content: center;
   border: 0 solid ${(props) => props.theme.colors.grey300};
   font-size: ${(props) => props.theme.fontSize.body2};
-  padding: 10px 30px;
+  padding: 10px 20px;
   cursor: pointer;
   box-shadow: 0 2px 12px rgba(25, 31, 40, 0.08);
   border-radius: 10px;
