+++ packages/gdsc-dju-admin/src/components/organisms/SideMenu/styled.ts
