+++ packages/gdsc-dju-admin/src/pages/Email/styled.ts
@@ -30,9 +30,6 @@ export const SelectedBoxSection = styled.section`
   row-gap: 10px;
   flex-direction: column;
   overflow: scroll;
-  width: 100%;
-  justify-content: center;
-  align-items: center;
   height: 100%;
   padding: 10px 20px;
   box-sizing: border-box;
@@ -42,27 +39,28 @@ export const TemplateEmailWrapper = styled.div`
   width: 300px;
 `;
 export const TemplateText = styled.div`
-  font-size: ${({ theme }) => theme.fontSize.h7};
+  font-size: ${({ theme }) => theme.fontSize.body1};
   color: ${({ theme }) => theme.colors.grey900};
   display: flex;
   align-items: center;
   justify-content: center;
   text-align: center;
-  width: 200px;
 `;
 export const EmailRightWrapper = styled.div`
   display: flex;
   align-items: center;
   flex-direction: column;
+
   box-sizing: border-box;
   width: 100%;
+  height: 100%;
 `;
 export const EmailRightInner = styled.div`
   padding: 0 20px;
   display: flex;
   box-sizing: border-box;
   width: 100%;
-  justify-content: flex-end;
+  height: 100%;
   flex-direction: column;
   align-items: center;
   flex-wrap: wrap;
@@ -70,17 +68,19 @@ export const EmailRightInner = styled.div`
 `;
 export const EmailLeftWrapper = styled.div`
   background: ${({ theme }) => theme.colors.grey50};
+  height: 100%;
 `;
 export const EmailLeftInner = styled.div`
   width: 240px;
   display: flex;
+  height: 100%;
   align-items: center;
-  justify-content: center;
   flex-direction: column;
   row-gap: 10px;
 `;
 export const EmailCategory = styled.div`
   font-size: ${({ theme }) => theme.fontSize.h7};
+  color: ${({ theme }) => theme.colors.grey900};
   display: flex;
   align-items: center;
   justify-content: center;
