+++ packages/gdsc-dju-admin/src/pages/EmailLog/index.tsx
@@ -1,21 +1,14 @@
 import { collection, limit, orderBy, query } from 'firebase/firestore';
-import React, { useRef } from 'react';
-import { GDSCButton } from '../../components/atoms/Button';
-import { TextInput } from '../../components/atoms/TextInput';
+import React from 'react';
 import EmailLogCard from '../../components/molecules/EmailLogCard';
 
 import { db } from '../../firebase/firebase';
 import { useFirestoreQuery } from '../../hooks/useFirebaseQuery';
 
 import { EmailLogTypeWithID } from '../../types/applicant';
-import { TemplateEmailWrapper, TemplateText } from '../Email/styled';
-import { LogWrapper, TemplateSelectWrapper } from './styled';
+import { LogWrapper } from './styled';
 
-const EmailLog: React.FC<{
-  template: string | null;
-  setTemplate: (template: string | null) => void;
-}> = ({ template, setTemplate }) => {
-  const templateRef = useRef<HTMLInputElement>(null);
+const EmailLog = () => {
   const emailLogQuery = query(
     collection(db, 'emailLogs'),
     orderBy('uploadDate', 'desc'),
@@ -25,27 +18,11 @@ const EmailLog: React.FC<{
   const emailLogs = useFirestoreQuery<EmailLogTypeWithID[]>(emailLogQuery);
 
   return (
-    <>
-      <TemplateSelectWrapper>
-        <TemplateText>
-          {template ? '템플릿이 없어요 :(' : '선택한 템플릿 '}
-          {template}
-        </TemplateText>
-        <TemplateEmailWrapper>
-          <TextInput ref={templateRef} placeholder={'템플릿을 입력해주세요.'} />
-        </TemplateEmailWrapper>
-        <GDSCButton
-          color={'blue900'}
-          text={'템플릿 선택'}
-          onClick={() => setTemplate(templateRef.current?.value ?? '')}
-        />
-      </TemplateSelectWrapper>
+    <LogWrapper>
       {emailLogs && emailLogs.length > 0 && (
-        <LogWrapper>
-          <EmailLogCard emailLogs={emailLogs} />
-        </LogWrapper>
+        <EmailLogCard emailLogs={emailLogs} />
       )}
-    </>
+    </LogWrapper>
   );
 };
