+++ packages/gdsc-dju-admin/src/pages/EmailLog/styled.ts
@@ -4,7 +4,6 @@ export const TemplateSelectWrapper = styled.div`
   display: flex;
   flex-direction: row;
   justify-content: flex-end;
-  height: 50px;
   flex-wrap: wrap;
   gap: 10px;
   width: 100%;
