+++ packages/gdsc-dju-admin/src/pages/index.tsx
@@ -1,6 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import AdminSidebar from '../components/organisms/AdminSidebar';
+import SideMenu from '../components/organisms/SideMenu';
 
 import AdminApplicants from './Applicants';
 import Email from './Email';
@@ -13,32 +13,20 @@ import {
 } from './styled';
 
 const Admin = () => {
-  const [template, setTemplate] = useState<string | null>(null);
-
   return (
-    <>
-      <AdminContainerWrapper>
-        <AdminSidebar />
-        <AdminContainer>
-          <AdminContainerInner>
-            <Routes>
-              <Route path={'/*'} element={<Home />} />
-              <Route path={'/recruit'} element={<AdminApplicants />} />
-              <Route path={'/email'} element={<Email template={template} />} />
-              <Route
-                path={'/email-log'}
-                element={
-                  <AdminEmailLog
-                    template={template}
-                    setTemplate={setTemplate}
-                  />
-                }
-              />
-            </Routes>
-          </AdminContainerInner>
-        </AdminContainer>
-      </AdminContainerWrapper>
-    </>
+    <AdminContainerWrapper>
+      <SideMenu />
+      <AdminContainer>
+        <AdminContainerInner>
+          <Routes>
+            <Route path={'/*'} element={<Home />} />
+            <Route path={'/recruit'} element={<AdminApplicants />} />
+            <Route path={'/email'} element={<Email />} />
+            <Route path={'/email-log'} element={<AdminEmailLog />} />
+          </Routes>
+        </AdminContainerInner>
+      </AdminContainer>
+    </AdminContainerWrapper>
   );
 };
